YoutubeChatBot/requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv~=0.21.0

# Machine Learning Utilities
numpy~=1.26.4
scikit-learn~=1.5.1
h11~=0.14.0
jupyter~=1.0.0
pip~=24.2
panel~=1.5.2
attrs~=23.1.0
zlib~=1.2.13
distro~=1.9.0
wheel~=0.44.0
pysocks~=1.7.1
protobuf~=4.25.3
openssl~=3.0.16
pyopenssl~=24.2.1
cryptography~=43.0.0
pillow~=10.4.0
docutils~=0.18.1
sphinx~=7.3.7
filelock~=3.13.1
pyasn1~=0.4.8
typing_extensions~=4.11.0
anyio~=4.2.0
sniffio~=1.3.0
pydantic~=2.8.2
httpx~=0.27.0
requests~=2.32.3
dask~=2024.8.2
rich~=13.7.1
colorama~=0.4.6
ipywidgets~=7.8.1
ipython~=8.27.0
notebook~=7.2.2
yaml~=0.2.5
pyyaml~=6.0.1
idna~=3.7
multidict~=6.0.4
pytest~=7.4.4
click~=8.1.7
pygments~=2.15.1
httpcore~=1.0.2
brotli~=1.0.9
zstandard~=0.23.0
charset-normalizer~=3.3.2
setuptools~=75.1.0
pytz~=2024.1
overrides~=7.4.0
cffi~=1.17.1
numba~=0.60.0
mypy~=1.11.2
regex~=2024.9.11
scipy~=1.13.1
threadpoolctl~=3.5.0
matplotlib~=3.9.2
blas~=1.0
mpmath~=1.3.0
sympy~=1.13.2
packaging~=24.1
platformdirs~=3.10.0
fsspec~=2024.6.1
snappy~=1.2.1
lz4~=4.3.2
distributed~=2024.8.2
aiohttp~=3.10.5
yarl~=1.11.0
paramiko~=2.8.1
ujson~=5.10.0
libarchive~=3.7.4
pyasn1-modules~=0.2.8
cachetools~=5.3.3
certifi~=2025.4.26
urllib3~=2.2.3
pandas~=2.2.2
tqdm~=4.66.5
joblib~=1.4.2
psutil~=5.9.0
cloudpickle~=3.0.0
aiosignal~=1.2.0
frozenlist~=1.4.0
aiohappyeyeballs~=2.4.0
pyarrow~=16.1.0
greenlet~=3.0.1
pyparsing~=3.1.2
annotated-types~=0.6.0
chardet~=4.0.0
tenacity~=8.2.3
tornado~=6.4.1
botocore~=1.34.69
sqlalchemy~=2.0.34
streamlit~=1.37.1
requests-toolbelt~=1.0.0
jinja2~=3.1.4
importlib-metadata~=7.0.1
pyodbc~=5.1.0
jsonpointer~=2.1
nltk~=3.9.1
beautifulsoup4~=4.12.3
gitpython~=3.1.43
h5py~=3.11.0
msgpack-python~=1.0.3
jsonpatch~=1.33
typer~=0.9.0
pydantic-settings~=2.6.1
tomli~=2.0.1
toml~=0.10.2
mypy_extensions~=1.0.0
networkx~=3.3
typing_extensions~=4.11.0
mypy_extensions~=1.0.0